No añadí nada nuevo solo estaba corrigiendo cositas como el llenado del modelo de beneficiario en el controller que hizo edwin.

+++ BankingApp.Infrastructure.Identity/Entities/ApplicationUser.cs
@@ -11,6 +11,6 @@ namespace BankingApp.Infrastructure.Identity.Entities
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public long CardIdentification { get; set; }
+        public int CardIdentification { get; set; }
     }
 }
